refactor: update MaterialSymbols folder and font imports
