+++ src/assets/fonts/material/MaterialSymbolsSharp.json
