+++ src/hooks/useCachedResources.ts
@@ -5,14 +5,13 @@ import { useEffect, useState } from "react";
 import { registerForPushNotificationsAsync } from "services/notifications";
 import { initializeApi } from "services/api";
 import * as SplashScreen from "expo-splash-screen";
-import MaterialSymbolsRounded from "assets/fonts/MaterialSymbolsRounded.ttf";
-import MaterialSymbolsOutlined from "assets/fonts/MaterialSymbolsOutlined.ttf";
-import MaterialSymbolsSharp from "assets/fonts/MaterialSymbolsSharp.ttf";
+import MaterialSymbolsRounded from "assets/fonts/material/MaterialSymbolsRounded.ttf";
+import MaterialSymbolsOutlined from "assets/fonts/material/MaterialSymbolsOutlined.ttf";
+import MaterialSymbolsSharp from "assets/fonts/material/MaterialSymbolsSharp.ttf";
 import { useCurrentUser } from "contexts/currentUserContext";
 import { useLanguage } from "contexts/languageContext";
 import { formattedLanguage } from "lib/formatters/languageFormatter";
 import * as Notifications from "expo-notifications";
-import SpaceMono from "../assets/fonts/SpaceMono-Regular.ttf";
 import GambarinoRegular from "../assets/fonts/Gambarino-Regular.ttf";
 import Inter400 from "../assets/fonts/inter/Inter-Regular.ttf";
 import Inter900 from "../assets/fonts/inter/Inter-Black.ttf";
@@ -68,7 +67,6 @@ export default function useCachedResources() {
         // Load fonts
         await Font.loadAsync({
           ...FontAwesome.font,
-          "space-mono": SpaceMono,
           "Gambarino-Regular": GambarinoRegular,
           Inter400,
           Inter900,
